pyproject.toml
[tool.poetry]
name = "package-python-api-flow"
version = "0.1.0"
description = "Package to enable to define flows to fetch and transform data"
authors = ["Dieik Ramos <dieik.ramos@syngenta.com>"]
readme = "README.md"
packages = [{include = "package_python_api_flow"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
pydantic = "^1.10.8"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
